requirements.txt
psycopg2-binary==2.8.2
Flask==1.0.3
Flask-RESTful==0.3.7
gunicorn==19.9.0
